requirements.txt
beautifulsoup4==4.12.3
groq==0.8.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
Requests==2.32.2
streamlit==1.35.0
uagents==0.12.2
yfinance==0.2.40
